Add files via upload
This nimBLE service allowed the muon detector to send a package of 512byte, with the service name MuonDetector able to be searched on NRF Connect. The main.c file has been renamed to avoid PR problems which may be caused by same names. This current branch project is based on a fixed set test data, needing further refinement.